Migrate debugger-related cases from main reducer
